2 bugs fixed:
Constructor: The top-left point of the vehicle is not added to the
hashmap if the row already exists in the hashmap, and if the vehicle is
vertical. Why: saw that the test failed when calling the top of a
vertical vehicle , tested in eclipse with the same file and saw that the
top of the verical vehicle was not being added where it should.

isGoal: Added a case where it returns false when the row does not exist in the HashMap instead of returning null. Why: Saw this because the test failed and said the method returned null instead of false.